replace file2chan by styxservers.  we now provide a file "data" wherever we are mounted.  we can now use e.g.:
	mount {cryptfile rawfile} /n/cryptfile
	mount -c {disk/kfs -r /n/cryptfile/data} /n/cryptkfs

to get a kfs on an encrypted disk image.

this cannot work with file2chan since there is no way to do a stat(5) on a file2chan, and hence kfs didn't know how much to write/read.

warning: there is no useful (secure) encryption yet.  the key is even hard-coded.  so don't use!